chore(model): RPC conversion code isolation (remaining)
