+++ crates/api-model/src/lib.rs
@@ -301,15 +301,6 @@ impl StateSla {
     }
 }
 
-impl From<StateSla> for rpc::forge::StateSla {
-    fn from(value: StateSla) -> Self {
-        rpc::forge::StateSla {
-            sla: value.sla.map(|sla| sla.into()),
-            time_in_state_above_sla: value.time_in_state_above_sla,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
