+++ crates/api-model/src/machine_boot_override.rs
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 use carbide_uuid::machine::MachineInterfaceId;
-use rpc::errors::RpcDataConversionError;
 use sqlx::postgres::PgRow;
 use sqlx::{FromRow, Row};
 
@@ -35,32 +34,5 @@ impl<'r> FromRow<'r, PgRow> for MachineBootOverride {
             custom_pxe: row.try_get("custom_pxe")?,
             custom_user_data: row.try_get("custom_user_data")?,
         })
-    }
-}
-
-impl TryFrom<rpc::forge::MachineBootOverride> for MachineBootOverride {
-    type Error = RpcDataConversionError;
-    fn try_from(value: rpc::forge::MachineBootOverride) -> Result<Self, Self::Error> {
-        let machine_interface_id =
-            value
-                .machine_interface_id
-                .ok_or(RpcDataConversionError::MissingArgument(
-                    "machine_interface_id",
-                ))?;
-        Ok(MachineBootOverride {
-            machine_interface_id,
-            custom_pxe: value.custom_pxe,
-            custom_user_data: value.custom_user_data,
-        })
-    }
-}
-
-impl From<MachineBootOverride> for rpc::forge::MachineBootOverride {
-    fn from(value: MachineBootOverride) -> Self {
-        rpc::forge::MachineBootOverride {
-            machine_interface_id: Some(value.machine_interface_id),
-            custom_pxe: value.custom_pxe,
-            custom_user_data: value.custom_user_data,
-        }
     }
 }
