+++ crates/api-model/src/nvl_logical_partition.rs
@@ -18,8 +18,6 @@
 use carbide_uuid::nvlink::NvLinkLogicalPartitionId;
 use chrono::{DateTime, Utc};
 use config_version::ConfigVersion;
-use rpc::errors::RpcDataConversionError;
-use rpc::forge as rpc_forge;
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::PgRow;
 use sqlx::{FromRow, Row};
@@ -32,74 +30,16 @@ pub struct NvLinkLogicalPartitionSearchFilter {
     pub name: Option<String>,
 }
 
-impl From<rpc_forge::NvLinkLogicalPartitionSearchFilter> for NvLinkLogicalPartitionSearchFilter {
-    fn from(filter: rpc_forge::NvLinkLogicalPartitionSearchFilter) -> Self {
-        NvLinkLogicalPartitionSearchFilter { name: filter.name }
-    }
-}
-
 #[derive(Debug, Clone)]
 pub struct NewLogicalPartition {
     pub id: NvLinkLogicalPartitionId,
     pub config: LogicalPartitionConfig,
 }
 
-impl TryFrom<rpc_forge::NvLinkLogicalPartitionCreationRequest> for NewLogicalPartition {
-    type Error = RpcDataConversionError;
-    fn try_from(
-        value: rpc_forge::NvLinkLogicalPartitionCreationRequest,
-    ) -> Result<Self, Self::Error> {
-        let id: NvLinkLogicalPartitionId = value.id.unwrap_or_else(|| uuid::Uuid::new_v4().into());
-
-        let conf = value.config.ok_or_else(|| {
-            RpcDataConversionError::InvalidArgument(
-                "NvLinkLogicalPartition config is empty".to_string(),
-            )
-        })?;
-
-        Ok(NewLogicalPartition {
-            id,
-            config: LogicalPartitionConfig::try_from(conf)?,
-        })
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Deserialize)]
 pub struct LogicalPartitionConfig {
     pub metadata: Metadata,
     pub tenant_organization_id: TenantOrganizationId,
-}
-
-impl TryFrom<rpc_forge::NvLinkLogicalPartitionConfig> for LogicalPartitionConfig {
-    type Error = RpcDataConversionError;
-
-    fn try_from(conf: rpc_forge::NvLinkLogicalPartitionConfig) -> Result<Self, Self::Error> {
-        if conf.tenant_organization_id.is_empty() {
-            return Err(RpcDataConversionError::InvalidArgument(
-                "NvLinkLogicalPartition organization_id is empty".to_string(),
-            ));
-        }
-
-        let tenant_organization_id =
-            TenantOrganizationId::try_from(conf.tenant_organization_id.clone()).map_err(|_| {
-                RpcDataConversionError::InvalidArgument(conf.tenant_organization_id)
-            })?;
-
-        Ok(LogicalPartitionConfig {
-            metadata: conf.metadata.unwrap_or_default().try_into()?,
-            tenant_organization_id,
-        })
-    }
-}
-
-impl TryFrom<LogicalPartitionConfig> for rpc_forge::NvLinkLogicalPartitionConfig {
-    type Error = RpcDataConversionError;
-    fn try_from(src: LogicalPartitionConfig) -> Result<Self, Self::Error> {
-        Ok(rpc_forge::NvLinkLogicalPartitionConfig {
-            metadata: Some(src.metadata.into()),
-            tenant_organization_id: src.tenant_organization_id.to_string(),
-        })
-    }
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
@@ -135,43 +75,6 @@ pub struct LogicalPartition {
 /// Returns whether a logical partition was deleted by user
 pub fn is_marked_as_deleted(partition: &LogicalPartition) -> bool {
     partition.deleted.is_some()
-}
-
-impl TryFrom<LogicalPartition> for rpc_forge::NvLinkLogicalPartition {
-    type Error = RpcDataConversionError;
-    fn try_from(src: LogicalPartition) -> Result<Self, Self::Error> {
-        let mut state = match &src.partition_state {
-            LogicalPartitionState::Provisioning => rpc_forge::TenantState::Provisioning,
-            LogicalPartitionState::Ready => rpc_forge::TenantState::Ready,
-            LogicalPartitionState::Error => rpc_forge::TenantState::Failed,
-            LogicalPartitionState::Deleting => rpc_forge::TenantState::Terminating,
-            LogicalPartitionState::Updating => rpc_forge::TenantState::Updating,
-        };
-
-        if is_marked_as_deleted(&src) {
-            state = rpc_forge::TenantState::Terminating;
-        }
-        let status = Some(rpc_forge::NvLinkLogicalPartitionStatus {
-            state: state as i32,
-        });
-
-        let config = rpc_forge::NvLinkLogicalPartitionConfig {
-            metadata: Some(rpc::Metadata {
-                name: src.name,
-                description: src.description,
-                ..Default::default()
-            }),
-            tenant_organization_id: src.tenant_organization_id.to_string(),
-        };
-
-        Ok(rpc_forge::NvLinkLogicalPartition {
-            id: Some(src.id),
-            config_version: src.config_version.version_string(),
-            status,
-            config: Some(config),
-            created: Some(src.created.into()),
-        })
-    }
 }
 
 #[derive(Debug, Serialize, Deserialize)]
