+++ crates/api-model/src/nvl_partition.rs
@@ -17,25 +17,16 @@
 
 use carbide_uuid::nvlink::{NvLinkDomainId, NvLinkLogicalPartitionId, NvLinkPartitionId};
 use chrono::{DateTime, Utc};
-use rpc::errors::RpcDataConversionError;
-use rpc::forge as rpc_forge;
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::PgRow;
 use sqlx::{FromRow, Row};
+
+use crate::errors::ModelError;
 
 #[derive(Clone, Debug, Default)]
 pub struct NvLinkPartitionSearchFilter {
     pub tenant_organization_id: Option<String>,
     pub name: Option<String>,
-}
-
-impl From<rpc_forge::NvLinkPartitionSearchFilter> for NvLinkPartitionSearchFilter {
-    fn from(filter: rpc_forge::NvLinkPartitionSearchFilter) -> Self {
-        NvLinkPartitionSearchFilter {
-            tenant_organization_id: filter.tenant_organization_id,
-            name: filter.name,
-        }
-    }
 }
 
 #[derive(Debug, Clone)]
@@ -62,7 +53,7 @@ impl NvlPartitionName {
 }
 
 impl TryFrom<String> for NvlPartitionName {
-    type Error = RpcDataConversionError;
+    type Error = ModelError;
     fn try_from(value: String) -> Result<Self, Self::Error> {
         Ok(NvlPartitionName(value))
     }
@@ -89,19 +80,6 @@ pub struct NvlPartition {
 /// Returns whether the NvLink partition was deleted
 pub fn is_marked_as_deleted(partition: &NvlPartition) -> bool {
     partition.deleted.is_some()
-}
-
-impl TryFrom<NvlPartition> for rpc_forge::NvLinkPartition {
-    type Error = RpcDataConversionError;
-    fn try_from(src: NvlPartition) -> Result<Self, Self::Error> {
-        Ok(rpc_forge::NvLinkPartition {
-            id: Some(src.id),
-            name: src.name.clone().into(),
-            nmx_m_id: src.nmx_m_id,
-            domain_uuid: Some(src.domain_uuid),
-            logical_partition_id: src.logical_partition_id,
-        })
-    }
 }
 
 #[derive(Debug, Serialize, Deserialize)]
