+++ crates/api-model/src/rpc_conv/nvl_partition.rs
@@ -1,0 +1,43 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use rpc::errors::RpcDataConversionError;
+use rpc::forge as rpc_forge;
+
+use crate::nvl_partition::{NvLinkPartitionSearchFilter, NvlPartition};
+
+impl From<rpc_forge::NvLinkPartitionSearchFilter> for NvLinkPartitionSearchFilter {
+    fn from(filter: rpc_forge::NvLinkPartitionSearchFilter) -> Self {
+        NvLinkPartitionSearchFilter {
+            tenant_organization_id: filter.tenant_organization_id,
+            name: filter.name,
+        }
+    }
+}
+
+impl TryFrom<NvlPartition> for rpc_forge::NvLinkPartition {
+    type Error = RpcDataConversionError;
+    fn try_from(src: NvlPartition) -> Result<Self, Self::Error> {
+        Ok(rpc_forge::NvLinkPartition {
+            id: Some(src.id),
+            name: src.name.clone().into(),
+            nmx_m_id: src.nmx_m_id,
+            domain_uuid: Some(src.domain_uuid),
+            logical_partition_id: src.logical_partition_id,
+        })
+    }
+}
